Avoid installing if already installed

+++ quick_start.sh
@@ -6,6 +6,7 @@ die () {
 }
 
 cd ~
+[ -d .shattings ] && die Shattings already installed
 curl -L https://github.com/darconeous/shattings/tarball/master | tar xvz || die Unable to download archive
 mv darconeous-shattings* .shattings || die Unable to move archive contents
 cd .shattings || die Unable to move into .shattings
